+++ src/utils/xss.js
@@ -49,6 +49,19 @@ xssUtils.stripTags = function (html, allowed) {
 };
 
 /**
+ * Remove console methods
+ * @private
+ * @param {string} html HTML in string form
+ * @returns {string} the modified value
+ */
+xssUtils.sanitizeConsoleMethods = function (html) {
+  const methods = ['assert', 'clear', 'count', 'debug', 'dirxml', 'dir', 'error', 'exception', 'groupCollapsed', 'groupEnd', 'group', 'info', 'log', 'markTimeline', 'profileEnd', 'profile', 'table', 'timeEnd', 'timeStamp', 'time', 'trace', 'warn'];
+  const expr = new RegExp(`console\\.(${methods.join('|')})((\\s+)?\\(([^)]+)\\);?)?`, 'igm');
+
+  return typeof html !== 'string' ? html : html.replace(expr, '');
+};
+
+/**
  * Remove Script tags and all onXXX functions
  * @private
  * @param {string} html HTML in string form
@@ -59,9 +72,7 @@ xssUtils.sanitizeHTML = function (html) {
   santizedHtml = santizedHtml.replace(/<[^>]+/g, match => match.replace(/(\/|\s)on\w+=(\'|")?[^"]*(\'|")?/g, '')); // eslint-disable-line
 
   // Remove console methods
-  const methods = ['assert', 'clear', 'count', 'debug', 'dirxml', 'dir', 'error', 'exception', 'groupCollapsed', 'groupEnd', 'group', 'info', 'log', 'markTimeline', 'profileEnd', 'profile', 'table', 'timeEnd', 'timeStamp', 'time', 'trace', 'warn'];
-  const expr = new RegExp(`console\\.(${methods.join('|')})((\\s+)?\\(([^)]+)\\);?)?`, 'igm');
-  santizedHtml = santizedHtml.replace(expr, '');
+  santizedHtml = this.sanitizeConsoleMethods(santizedHtml);
 
   // Remove nested script tags
   santizedHtml = santizedHtml.replace(/<\/script>/g, '');