+++ docs/CHANGELOG.md
@@ -7,9 +7,8 @@
 ### v4.19.0 Features
 
 ### v4.19.0 Fixes
-- `[Datagrid]` Fixed an issue for xss where console.log was not sanitizing and make grid to not render. ([#1941](https://github.com/infor-design/enterprise/issues/1941))
 
-### v4.19.0 Chore & Maintenance
+- `[Datagrid]` Fixed an issue for xss where console.log was not sanitizing and make grid to not render. ([#1941](https://github.com/infor-design/enterprise/issues/1941))
 
 ### v4.19.0 Chores & Maintenance